REFACTORING: Outliner component file split by subcomponents.
